requirements.txt
alembic==1.0.10
amqp==2.4.2
aniso8601==6.0.0
anyjson==0.3.3
atomicwrites==1.3.0
attrs==19.1.0
billiard==3.6.0.0
celery==4.3.0
Click==7.0
coverage==4.5.3
Flask==1.0.3
Flask-Cors==3.0.7
Flask-JWT-Extended==3.18.2
Flask-Migrate==2.5.2
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.0
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
kombu==4.5.0
Mako==1.0.12
MarkupSafe==1.1.1
marshmallow==3.0.0rc6
more-itertools==7.0.0
pluggy==0.11.0
psycopg2==2.8.3
psycopg2-binary==2.8.3
py==1.8.0
PyJWT==1.7.1
pytest==4.5.0
pytest-cov==2.7.1
python-dateutil==2.8.0
python-dotenv==0.10.2
python-editor==1.0.4
pytz==2019.1
six==1.12.0
SQLAlchemy==1.3.4
vine==1.3.0
wcwidth==0.1.7
Werkzeug==0.15.4
